# tequilapi/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this documentation is to provide developers an insight of how to\ninteract with Mysterium Node via Tequila API.\nThis should demonstrate all the possible API calls with described parameters and responses.",
    "title": "Tequila API",
    "version": "dev"
  },
  "host": "127.0.0.1:4050",
  "paths": {
...
    "/terms": {
      "get": {
        "description": "Return an object with the current terms config",
        "tags": [
          "Terms"
        ],
        "summary": "Get terms",
        "operationId": "getTerms",
        "responses": {
          "200": {
            "description": "Terms object",
            "schema": {
              "$ref": "#/definitions/TermsResponse"
            }
          }
        }
      },
      "post": {
        "description": "Takes the given data and tries to update terms agreement config.",
        "tags": [
          "Terms"
        ],
        "summary": "Update terms agreement",
        "operationId": "updateTerms",
        "parameters": [
          {
            "description": "Required data to update terms",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/TermsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Terms agreement updated"
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "APIError": {
      "type": "object",
      "title": "APIError represents an error response from REST API service.",
      "properties": {
        "error": {
          "$ref": "#/definitions/Err"
        },
        "path": {
          "type": "string",
          "x-go-name": "Path"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "apiErrorSwagger",
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
...
    "Err": {
      "type": "object",
      "title": "Err contains the error details.",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "Code"
        },
        "detail": {
          "type": "string",
          "x-go-name": "Detail"
        },
        "fields": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/FieldError"
          },
          "x-go-name": "Fields"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
...
    "FieldError": {
      "type": "object",
      "title": "FieldError contains the reason why a field failed validation.",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "Code"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
...
    "TermsRequest": {
      "type": "object",
      "title": "TermsRequest object is accepted by terms endpoints.",
      "properties": {
        "agreed_consumer": {
          "type": "boolean",
          "x-go-name": "AgreedConsumer",
          "example": false
        },
        "agreed_provider": {
          "type": "boolean",
          "x-go-name": "AgreedProvider",
          "example": false
        },
        "agreed_version": {
          "type": "string",
          "x-go-name": "AgreedVersion",
          "example": "0.0.27"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "TermsResponse": {
      "type": "object",
      "title": "TermsResponse object is returned by terms endpoints.",
      "properties": {
        "agreed_consumer": {
          "type": "boolean",
          "x-go-name": "AgreedConsumer",
          "example": false
        },
        "agreed_provider": {
          "type": "boolean",
          "x-go-name": "AgreedProvider",
          "example": false
        },
        "agreed_version": {
          "type": "string",
          "x-go-name": "AgreedVersion",
          "example": "0.0.27"
        },
        "current_version": {
          "type": "string",
          "x-go-name": "CurrentVersion",
          "example": "0.0.27"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
  }
}